# backend/src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: Pokemon Center Api
  description: Pokemon Center Api
  version: 1.0.0
servers:
  - url: /
paths:
...
  /cadastro/pessoa:
    post:
      summary: Realiza o cadastro de uma pessoas no banco de dados.
      operationId: cadastrarPessoa
      tags:
        - Cadastro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pessoa'
      responses:
        '200':
          description: Sucesso ao realizar cadastro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mensagem'
        '500':
          description: Erro ao realizar cadastro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mensagem'
  /cadastro/funcionario:
    post:
      summary: Realiza o cadastro de um funcionário no banco de dados.
      operationId: cadastrarFuncionario
      tags:
        - Cadastro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/funcionario'
      responses:
        '200':
          description: Sucesso ao realizar cadastro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mensagem'
        '500':
          description: Erro ao realizar cadastro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mensagem'
  /cadastro/medico:
    post:
      summary: Realiza o cadastro de um médico no banco de dados.
      operationId: cadastrarMedico
      tags:
        - Cadastro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/medico'
      responses:
        '200':
          description: Sucesso ao realizar cadastro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mensagem'
        '500':
          description: Erro ao realizar cadastro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mensagem'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  # 2) Apply the security globally to all operations
  schemas:
    mensagem:
      type: object
      properties:
        mensagem:
          type: string
          example: "descricão"
    pessoa:
      type: object
      properties:
        cpf:
          type: string
          example: "15984533"
        nome:
          type: string
          example: "Gabriel Arrighi"
        email:
          type: string
          example: "arrighi@gmail.com"
        telefone:
          type: string
          example: 38999225757
        cep:
          type: string
          example: 30190002
        lougadouro:
          type: string
          example: "Avenida Augusto de Lima"
        bairro:
          type: string
          example: "Barro Preto"
        cidade:
          type: string
          example: "Belo Horizonte"
        estado:
          type: string
          example: "Minas Gerais"
    funcionario:
      type: object
      properties:
        cpf:
          type: string
          example: "15984533"
        nome:
          type: string
          example: "Gabriel Arrighi"
        email:
          type: string
          example: "arrighi@gmail.com"
        telefone:
          type: string
          example: 38999225757
        cep:
          type: string
          example: 30190002
        lougadouro:
          type: string
          example: "Avenida Augusto de Lima"
        bairro:
          type: string
          example: "Barro Preto"
        cidade:
          type: string
          example: "Belo Horizonte"
        estado:
          type: string
          example: "Minas Gerais"
        dataContrato:
          type: string
          example: "Avenida Augusto de Lima"
        salario:
          type: number
          example: 15.000,00
        senha:
          type: string
          example: "Belo Horizonte"
    medico:
      type: object
      properties:
        cpf:
          type: string
          example: "15984533"
        nome:
          type: string
          example: "Gabriel Arrighi"
        email:
          type: string
          example: "arrighi@gmail.com"
        telefone:
          type: string
          example: 38999225757
        cep:
          type: string
          example: 30190002
        lougadouro:
          type: string
          example: "Avenida Augusto de Lima"
        bairro:
          type: string
          example: "Barro Preto"
        cidade:
          type: string
          example: "Belo Horizonte"
        estado:
          type: string
          example: "Minas Gerais"
        dataContrato:
          type: string
          example: "Avenida Augusto de Lima"
        salario:
          type: number
          example: 15.000,00
        senha:
          type: string
          example: "Belo Horizonte"
        especialidade:
          type: array
          items:
            $ref: '#/components/schemas/especialidade'
        crm:
          type: string
          example: "15555"
        codigo:
          type: string
          example: "15984533"
    especialidade:
      type: string
      example: "ortopedista"
...